// generated/src/app/api-models/language-master.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class LanguageMaster {
  languageCode: string;
  description: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  languageId: number;
  oracleLanguageCode: string;
}